fixup! AmqpForwardAttribute mailet: ignore errors when creating an exchange that already exists with a different configuration.
- add test cases
